Share remaining items except activity
