+++ packages/backend/src/config/config.local.ts
@@ -1,27 +1,11 @@
 import { Env } from '@l2beat/backend-tools'
-import { chains, tokenList } from '@l2beat/config'
-import { ChainId, UnixTime } from '@l2beat/shared-pure'
+import { UnixTime } from '@l2beat/shared-pure'
 
 import { Config } from './Config'
-import { getChainDiscoveryConfig } from './getChainDiscoveryConfig'
-import { getChainsWithTokens } from './getChainsWithTokens'
-import { getChainTvlConfig } from './getChainTvlConfig'
 import { makeConfig } from './makeConfig'
 
 export function getLocalConfig(env: Env): Config {
   const activityEnabled = env.boolean('ACTIVITY_ENABLED', false)
-  const activityProjectsExcludedFromApi = env.optionalString(
-    'ACTIVITY_PROJECTS_EXCLUDED_FROM_API',
-  )
-  const livenessEnabled = env.boolean('LIVENESS_ENABLED', false)
-  const updateMonitorEnabled = env.boolean('WATCHMODE_ENABLED', false)
-  const diffHistoryEnabled = env.boolean('DIFF_HISTORY_ENABLED', false)
-  const discordToken = env.optionalString('DISCORD_TOKEN')
-  const internalDiscordChannelId = env.optionalString(
-    'INTERNAL_DISCORD_CHANNEL_ID',
-  )
-  const discordEnabled = !!discordToken && !!internalDiscordChannelId
-  const finalityEnabled = env.boolean('FINALITY_ENABLED', false)
 
   // TODO: This should probably be configurable
   const minTimestamp = UnixTime.now().add(-7, 'days').toStartOf('hour')
@@ -32,28 +16,13 @@ export function getLocalConfig(env: Env): Config {
       isLocal: true,
       minTimestampOverride: minTimestamp,
     }),
-    liveness: livenessEnabled && {
-      bigQuery: {
-        clientEmail: env.string('LIVENESS_CLIENT_EMAIL'),
-        privateKey: env.string('LIVENESS_PRIVATE_KEY').replace(/\\n/g, '\n'),
-        projectId: env.string('LIVENESS_PROJECT_ID'),
-        queryLimitGb: env.integer('LIVENESS_BIGQUERY_LIMIT_GB', 15),
-        queryWarningLimitGb: env.integer(
-          'LIVENESS_BIGQUERY_WARNING_LIMIT_GB',
-          8,
-        ),
-      },
-      // TODO: figure out how to set it for local development
-      minTimestamp: UnixTime.fromDate(new Date('2023-05-01T00:00:00Z')),
-    },
-    finality: finalityEnabled,
     activity: activityEnabled && {
       starkexApiKey: env.string('STARKEX_API_KEY'),
       starkexCallsPerMinute: env.integer('STARKEX_CALLS_PER_MINUTE', 600),
       skipExplicitExclusion: true,
-      projectsExcludedFromAPI: activityProjectsExcludedFromApi
-        ? activityProjectsExcludedFromApi.split(' ')
-        : [],
+      projectsExcludedFromAPI:
+        env.optionalString('ACTIVITY_PROJECTS_EXCLUDED_FROM_API')?.split(' ') ??
+        [],
 
       projects: {
         ethereum: {
@@ -127,30 +96,5 @@ export function getLocalConfig(env: Env): Config {
         },
       },
     },
-    statusEnabled: env.boolean('STATUS_ENABLED', true),
-    updateMonitor: updateMonitorEnabled && {
-      runOnStart: env.boolean('UPDATE_MONITOR_RUN_ON_START', true),
-      discord: discordEnabled && {
-        token: discordToken,
-        publicChannelId: env.optionalString('PUBLIC_DISCORD_CHANNEL_ID'),
-        internalChannelId: internalDiscordChannelId,
-        callsPerMinute: 3000,
-      },
-      chains: [
-        getChainDiscoveryConfig(env, 'ethereum'),
-        getChainDiscoveryConfig(env, 'arbitrum'),
-        getChainDiscoveryConfig(env, 'bsc'),
-        getChainDiscoveryConfig(env, 'celo'),
-        getChainDiscoveryConfig(env, 'gnosis'),
-        getChainDiscoveryConfig(env, 'linea'),
-        getChainDiscoveryConfig(env, 'optimism'),
-        getChainDiscoveryConfig(env, 'polygonpos'),
-        getChainDiscoveryConfig(env, 'polygonzkevm'),
-      ],
-    },
-    diffHistory: diffHistoryEnabled && {
-      chains: [getChainDiscoveryConfig(env, 'ethereum')],
-    },
-    chains: chains.map((x) => ({ name: x.name, chainId: ChainId(x.chainId) })),
   }
 }
